+++ lib/MongoDB/Collection.js
@@ -10,15 +10,15 @@ const stripId = Symbol('stripId');
 class Collection {
     /**
      * @param {import('mongodb').Db} db
-     * @param {import('mongodb').MongoClient} mongoClient
-     * @param {import('../const/collections').MongoClient} collection
+     * @param {import('mongodb').MongoClient} client
+     * @param {import('./collections')} collection
      */
-    constructor(db, mongoClient, collection) {
+    constructor(db, client, collection) {
         /** @type {import('mongodb').Collection<T>} */
         this.index = collection.index;
         this.entityName = collection.entityName;
         this.collection = db.collection(collection.name);
-        this._client = mongoClient;
+        this._client = client;
     }
 
     [validateId](id) {
@@ -45,6 +45,12 @@ class Collection {
         await this.collection.createIndex({ [name]: 1 }, { name, unique });
     }
 
+    /**
+     * @template T
+     * @param {T} options
+     * @param {T} settings
+     * @returns {T}
+     */
     async create(options, { applyId = false } = {}) {
         try {
             const response = await this.collection.insertOne({ ...options });
@@ -58,6 +64,12 @@ class Collection {
         }
     }
 
+    /**
+     * @template T
+     * @param {T} options
+     * @param {T} settings
+     * @returns {T}
+     */
     async createMany(options, { ordered = false } = {}) {
         try {
             const result = await this.collection.insertMany(options, { ordered });
