+++ lib/MongoDB/DataSources.js
@@ -2,7 +2,7 @@ const { ObjectID } = require('mongodb');
 const errors = require('../errors');
 const Collection = require('./Collection');
 const { stripId } = require('./Collection');
-const collections = require('../const/collections');
+const collections = require('./collections');
 
 /**
  * @typedef {import('../DataSource').DataSourcesInterface} DataSourcesInterface
