+++ lib/MongoDB/Jobs.js
@@ -1,5 +1,5 @@
 const Collection = require('./Collection');
-const collections = require('../const/collections');
+const collections = require('./collections');
 
 class Jobs extends Collection {
     constructor(db, client) {
