+++ lib/MongoDB/TensorBoards.js
@@ -1,5 +1,5 @@
 const Collection = require('./Collection');
-const collections = require('../const/collections');
+const collections = require('./collections');
 
 class TensorBoards extends Collection {
     constructor(db, client) {
