+++ lib/MongoDB/algorithms/store.js
@@ -1,7 +1,7 @@
 const Collection = require('../Collection');
 const Versions = require('./versions');
 const Builds = require('./builds');
-const collections = require('../../const/collections');
+const collections = require('../collections');
 
 class Algorithms extends Collection {
     constructor(db, client) {
