+++ lib/MongoDB/algorithms/versions.js
@@ -1,5 +1,5 @@
 const Collection = require('../Collection');
-const collections = require('../../const/collections');
+const collections = require('../collections');
 
 class Versions extends Collection {
     constructor(db, client) {
