+++ lib/MongoDB/index.js
@@ -10,7 +10,7 @@ const TensorBoards = require('./TensorBoards');
 const WebhookStatus = require('./webhooks/Status');
 const WebhookResult = require('./webhooks/Result');
 
-/** @typedef {import('./../types').MongoConfig} MongoConfig */
+/** @typedef {import('../types').MongoConfig} MongoConfig */
 /** @typedef {import('../Provider').ProviderInterface} ProviderInterface */
 
 /** @implements {ProviderInterface} */
