+++ lib/MongoDB/readme/Algorithms.js
@@ -1,5 +1,5 @@
 const Collection = require('../Collection');
-const collections = require('../../const/collections');
+const collections = require('../collections');
 
 class Algorithms extends Collection {
     constructor(db, client) {
