+++ lib/MongoDB/webhooks/Status.js
@@ -1,5 +1,5 @@
 const Collection = require('../Collection');
-const collections = require('../../const/collections');
+const collections = require('../collections');
 
 class WebhookStatus extends Collection {
     constructor(db, client) {
